@@ -1,6 +1,6 @@
 {
-    "name": "horde/listheaders",
-    "description": "List headers parsing library",
+    "name": "bytestream/horde-listheaders",
+    "description": "Horde list headers parsing library",
     "type": "library",
     "homepage": "https://www.horde.org/libraries/Horde_ListHeaders",
     "license": "LGPL-2.1",
@@ -11,31 +11,23 @@
             "role": "lead"
         }
     ],
-    "version": "1.2.6",
-    "time": "2017-11-11",
-    "repositories": [
-        {
-            "type": "pear",
-            "url": "https://pear.horde.org"
-        }
-    ],
     "require": {
-        "php": "^5.3 || ^7",
-        "pear-pear.horde.org/Horde_Mail": "^2",
-        "pear-pear.horde.org/Horde_Translation": "^2.2",
-        "pear-pear.horde.org/Horde_Util": "^2"
+        "php": "^7.4 || ^8.0",
+        "bytestream/horde-mail": "dev-master",
+        "bytestream/horde-translation": "^2.2",
+        "bytestream/horde-util": "^2"
     },
-    "suggest": {
-        "pear-pear.horde.org/Horde_Mime": "^2.5",
-        "pear-pear.horde.org/Horde_Test": "^2.1"
-    },
-    "replace": {
-        "pear-pear.horde.org/Horde_ListHeaders": "1.*",
-        "pear-horde/Horde_ListHeaders": "1.*"
+    "require-dev": {
+        "phpunit/phpunit": "^9.5"
     },
     "autoload": {
         "psr-0": {
             "Horde_ListHeaders": "lib/"
+        }
+    },
+    "autoload-dev": {
+        "psr-0": {
+            "Horde_ListHeaders": "test/"
         }
     }
 }
